created a seperate file to treat database-server interaction. Currently only send works.
