--- LodeLMmvn/src/main/java/utils/FirebaseInit.java
@@ -1,26 +1,0 @@
-package utils;
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import com.google.firebase.database.FirebaseDatabase;
-import java.io.FileInputStream;
-import java.io.IOException;
-
-public class FirebaseInit {
-
-    public static void initialize() {
-        try {
-            FileInputStream serviceAccount = new FileInputStream("/Users/Aniku/Downloads/serviceAccountKey.json");
-
-            FirebaseOptions options = new FirebaseOptions.Builder()
-                .setCredentials(GoogleCredentials.fromStream(serviceAccount))
-                .setDatabaseUrl("https://lodelm-936f0-default-rtdb.firebaseio.com/")
-                .build();
-
-            FirebaseApp.initializeApp(options);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
